feat(wrapper):[TRI-810] Rework Helm charts
IRS Helm chart now only contains IRS and has dependencies on the stock charts of Minio, Grafana and Prometheus.

Grafana and Prometheus are optional and can be disabled.

A new Helm chart is available which provides the EDC consumer.

The IRS local chart now depends on those two charts and includes them locally.