feat(registryclient):[TRI-1470] Improve registry client library
Added new method to retrieve globalAssetIds for a BPN directly. This just uses the existing methods.

Refactored the existing code to be more comprehensible and added some log output.
The EndpointDataReference is now renewed on demand if the token expires during the runtime.
Renamed DigitalTwinRegistryService#lookupShells to lookupShellIdentifiers to be more precise. The old method points to the new one and is marked as deprecated now.